Inclusão da ENV

+++ src/modules/projects/services/DeleteProjectService.ts
@@ -1,6 +1,6 @@
 import AppError from '@shared/errors/AppError';
 import IProjectsRepository from '@modules/projects/repositories/IProjectsRepository';
-// import Project from '@modules/projects/infra/typeorm/entities/Project';
+import { validate } from 'uuid';
 import { inject, injectable } from 'tsyringe';
 
 @injectable()
@@ -10,12 +10,28 @@ class DeleteProjectService {
     private projectRepository: IProjectsRepository,
   ) {}
 
-  public async execute(project_id: string): Promise<void> {
+  public async execute(project_id: string): Promise<boolean> {
     if (!project_id) {
       throw new AppError('Project ID is null');
     }
 
+    const checkProjectIdisUuid = validate(project_id);
+
+    if (!checkProjectIdisUuid) {
+      throw new AppError('Project ID is not validy');
+    }
+
+    const checkExistsProject = await this.projectRepository.findById(
+      project_id,
+    );
+
+    if (!checkExistsProject) {
+      throw new AppError('Project ID not found');
+    }
+
     await this.projectRepository.delete(project_id);
+
+    return true;
   }
 }
